Split trigger mapper from form mapper to entity scope
