+++ .gitignore
@@ -54,3 +54,4 @@ com_crashlytics_export_strings.xml.DS_Store
 out
 bin
 gen
+.DS_Store
