+++ core/src/main/java/com/jaspersoft/android/sdk/service/report/schedule/BaseTriggerMapper.java
@@ -5,6 +5,7 @@ import com.jaspersoft.android.sdk.network.entity.schedule.JobTriggerEntity;
 import com.jaspersoft.android.sdk.service.data.schedule.JobForm;
 import com.jaspersoft.android.sdk.service.data.schedule.UntilEndDate;
 
+import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Locale;
@@ -34,6 +35,23 @@ abstract class BaseTriggerMapper {
         triggerEntity.setTimezone(timeZone.getID());
     }
 
+    protected void mapCommonEntityTriggerFieldsOnEntity(JobForm.Builder form, JobFormEntity entity) {
+        JobTriggerEntity triggerEntity = entity.getTrigger();
+        String startDate = triggerEntity.getStartDate();
+        if (startDate != null) {
+            Date date = parseDate(startDate);
+            if (date != null) {
+                form.withStartDate(date);
+            }
+        }
+
+        String timezone = triggerEntity.getTimezone();
+        if (timezone != null) {
+            TimeZone timeZone = TimeZone.getTimeZone(timezone);
+            form.withTimeZone(timeZone);
+        }
+    }
+
     protected void mapEndDate(UntilEndDate endDate, JobTriggerEntity triggerEntity) {
         UntilEndDate date = endDate;
 
@@ -43,9 +61,19 @@ abstract class BaseTriggerMapper {
         triggerEntity.setEndDate(untilDate);
     }
 
+    protected Date parseDate(String target) {
+        try {
+            return DATE_FORMAT.parse(target);
+        } catch (ParseException e) {
+            return null;
+        }
+    }
+
     protected SimpleDateFormat getCommandDateFormat() {
         return DATE_FORMAT;
     }
 
     public abstract void toTriggerEntity(JobForm form, JobFormEntity entity);
+
+    public abstract void toDataForm(JobForm.Builder form, JobFormEntity entity);
 }
